+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added optional `method` field to send DTMF properties. Supported values are
   `"sip-info"`, `"telephone-event"`, and `"auto"`.
 
+- Added optional `digitDurationMs` field to send DTMF properties.
+
 ## [0.24.0] - 2026-03-09
 
 ### Added